chore: moved formatNotional to util file
